+++ atelier/src/metrics/mod.rs
@@ -1,1 +1,2 @@
 pub mod market;
+pub mod orderbook;
